20190821-pyinstaller-dl/requirements.txt
h5py==2.9.0
html5lib==0.9999999
Keras==2.1.5
Markdown==3.1.1
numpy==1.15.3
Pillow==4.1.1
protobuf==3.9.1
PyInstaller==3.3
pywin32==224
PyYAML==5.1.2
requests==2.22.0
scipy==1.3.1
six==1.12.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
wxPython==4.0.6
